requirements.txt
-i https://pypi.org/simple
aiosmtpd==1.4.4.post2; python_version ~= '3.7'
alembic==1.13.0; python_version >= '3.8'
atpublic==4.0; python_version >= '3.8'
attrs==23.1.0; python_version >= '3.7'
babel==2.14.0; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
email-validator==2.1.0.post1; python_version >= '3.8'
flask==3.0.0; python_version >= '3.8'
flask-babel==4.0.0; python_version >= '3.8' and python_version < '4.0'
flask-login==0.6.3; python_version >= '3.7'
flask-mail==0.9.1
flask-migrate==4.0.5; python_version >= '3.6'
flask-moment==1.0.5; python_version >= '3.6'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
flask-wtf==1.2.1; python_version >= '3.8'
greenlet==3.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.6; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
langdetect==1.0.9
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.23; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
wtforms==3.1.1; python_version >= '3.8'
